requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
cloudpickle==2.2.0
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
gym==0.26.2
gym-notices==0.0.8
highway-env==1.7
idna==3.4
importlib-metadata==5.0.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.12
zipp==3.10.0
torchsummary==1.5.1
tqdm==4.64.1